run consistency check against a specific full backup when many fulls are in same path ()
run consistency check against specific full backup

The path supplied to a consistency check (`neo4j-admin database check
—from-path`)
may now also specify a particular full backup in a backup directory,
rather than the directory itself. The effect will be to check that the
particular backup is consistent. This may not be the most recent backup.

Also allow an individual dump file to be specified in the same way.

changelog:
* add the ability to run consistency check against a specific full
backup when many fulls are in same path
* allow consistency checking of a dump to check a specific dump file