fix: planning nested existence queries  ()
For deeply nested EXIST queries (example: macro benchmark graphql ->
Query_DeeplyNestedWithRelationshipFilter), we would sometimes notice
that the existence query was executed immediately after the expand
instead of a has labels filter.
This would present us with two problems:
1. We are potentially missing out on a cheap filter on multiple rows
before running existence queries
2. If the existence query depends on the unbound node in the expand, we
are missing out on the label information on this node.

While the regression itself was caused by making the
RemoteBatchPropertiesWithFilter costlier, the underlying problem was
that we were underestimating the cardinality of a few expands (we
consider them selective while they are not). By planning the has labels
before the existence query, we can ensure that the label information is
propagated into the subquery planner.
A side-effect of this change: For the graphql query, one should also see
an improvement in planning times. Instead of 1159 plan comparisons, we
are now down to 684.

This commit will also introduce previouslyCachedProperties as the cache
key for the planning of nested subqueries. As a result we can ensure
that cached properties on the LHS are propagated through the RHS of
semiApply to avoid redundant remoteBatchProperties