Add FABRIC capability to Intra Bolt Connector ()
This capability is defined in the Bolt Server, but this is not exposed
at all in none of the connectors.

With the implementation of the new GQL messages for failures, all the
messages will be changed when Bolt 6.0 is negotiated. This is can lead
to the legacy message being gone when query is routed over composite.

For solving this, the Bolt server is sending the legacy message in the
diagnostic record when FABRIC capability is set. However, this
capability is not being offered during the handshake.