Expose vectored writes in PageSwapper using the same format as the page cache. ()
Previously, swappers did not provide an asynchronous vectored write API.
The existing method in AsyncBlockSwapper always required fresh buffers
to be passed in from the outside. In practice, however, real users (such
as checkpointing) reuse buffers, where the buffer size and the number of
pages/buffers to flush are controlled through additional counters.

This PR introduces vectored writes but does not yet utilize page refs
array or page stumps passed arrays. That will be updated to be used in
following work.