Better memory tracking for index population ()
Previously it would register memory allocations for its internal kernel
cursors, and not reporting releasing that memory! This was a problem for
any cursor that registered memory allocation, e.g.
BlockRelationshipTraversalCursor and RecordPropertyCursor, resulting in
a memory _reporting_ leak, of sorts.

Now, the majority of allocated memory by far is that of the values to
index when batching them up. This commit shifts so that the values to be
indexed are properly memory tracked during population whereas the cursor
memory allocations aren't.