allow injection of `startTimeMillis` into `KernelTransactionImplementation.initialize()` and the `VirtualKernelTransaction` constructor ()
This PR proposes being able to inject `startTimeMillis` to
`KernelTransactionImplementation.initialize()` and the
`VirtualKernelTransaction` constructor, rather than setting it
internally.

The reason for this is that In order to facilitate Attribute-Based
Access Control (ABAC), we need to synchronize the value used for
transaction start time with the time used to evaluate time-based ABAC
rules.

With ABAC, roles can be granted based on attributes of the user or the
environment (e.g. time). This could include something like:

_grant the 'nightUsers' role between the hours of 19:00 and 07:00_

This means that we need the value used for transaction `startTime` to be
precisely the same as the time used to evaluate ABAC rules.

Rules like these need to be evaluated _before_ transactions are started
to make sure that the user is indeed authorized to even start a
transaction (just like for existing permission checks). So we need to be
able to capture a single time which will be used for `startTime` and the
the time operands of rules like the one above.

Follows on from [this
conversation](https://neo4j.slack.com/archives/C02JJGEPD/p1755076866775589).