Separate parts not backed by store from MetadataStores
For legacy reasons the MetadataStore has been responsible for some transaction tracking that is not
stored on disk.
Starting now MetadataStore is only responsible for the parts that are backed by an actual store file (store ids).
All other things (that were kept solely in memory)  is separated out to the new LogMetadataProvider.
The MetadataCache has been combined with the new LogMetadataProvider to a single provider since they
are both initialized based on values from the transaction logs on startup.