Fix Cypher 25 bug with dropping composite database with aliases ()
When dropping composite database with constituent aliases using
parameters for the database name in Cypher 25 the aliases neither
resulted in an error (default behaviour) nor got dropped (with cascade
aliases). Instead they got left as ghost aliases that didn't show up in
`SHOW ALIASES` but still blocked the creation of new aliases with the
same full name. Luckily ghost aliases can still be dropped.

The error scenario is quite specific, you need a composite database with
a `.` in the name, have constituent aliases in it and try to drop it
using a parameter in Cypher 25 so hopefully not too many are actually
impacted by this.