Handle TransactionLog startup inside catchup protocol on enveloped files when not aligned on tx start ()
In `RawTxLogCatchUpWriter` a `TransactionLogFile` is initialized for
writing inside `TxLogCatchupSession.createIfNotExists()` and as the
context is relatively isolated there is no proper last closed
transaction position to give and instead the file start is used. If that
file starts with the a partial enveloped entry it is not possible to
start parsing entries immediately after the header and
`EnvelopedReadChannel.alignWithStartEntry` needs to be called. This
change prevents the parsing exceptions that were occurring and correctly
finds the point to start appending.

I have also fixed the DumpLogicalTool which similarly threw an exception
if the logs were pruned to this situation and thus was unable to give
information on correctly formed log files