Eagerness Analysis - Merging of candidate lists ()
While going through the list of candidate lists, try to merge with
multiple candidate lists that are in the buffer instead of only one
Old behavior: ({1,[2]}, {1,[3]}, {[1]}) => ({[1]}, {1,[3]})
New behavior: ({1,[2]}, {1,[3]}, {[1]}) => ({[1]})
The first two cannot be merged, so both end up in the buffer.
The third can be merged with both candidate lists in the buffer.