Introduce `application/vnd.neo4j.query.v1.0` in Query Api ()
The Query API typed json needs to be versioned to add new types. For
consistency, the `application/vnd.neo4j.query` is defined as the version
`1.0` of the MimeType and aliased as
`application/vnd.neo4j.query.v1.0`.

Both mime types are the same and the original mime type won't be
removed.