Page prefetcher array out of bounds fix ()
Fix bug in `PagedFile::touch` where we can get
`ArrayIndexOutOfBoundsException` when accessing translation table.
If touching page concurrently with the pin that grows file *and*
translation table there is period of time when last page is already
updated but translation table is not adjusted yet.
Also ensure that latches are correctly released even if there is failure
inside `grabPageFaultLatches`