Index population does not keep horizon until it started scanning  ()
We have limited number of population threads and when there are number
of submitted population jobs is greater then available population
threads, some jobs will be pending and holding global horizon equal to
`transactionIdCreatedIndexes`.
That can produce a deadlock with population jobs that are running and
waiting until horizon is progressed at the end of the population. Thus
jobs that are running can't progress, and jobs that are pending and
holding horizon can't start.

To prevent this, population job starts with `Long.MAX_VALUE` horizon,
which is updated to the real value before the store scan, when the job
is already running.
Also `UnboundedReadVersionContext` carries `oldestTransactionId` now,
because that context is used to do writes in the versioned tree and is
needed to prevent tree history from touching recycled pages.