Add some more special cases for the check on general Cypher against system database ()
Should, for example, print out `UNION` as an invalid clause. Keeps
traversing to find more invalid clauses as they are very top level and
can easily contain more invalid things we'd want to error on.

Also updates the message to have distinct clauses instead of repeating
them if they appear more than once in the query (as was done for show
commands not allowed on system)

And other unrelated minor changes I noticed as part of the composable
show command work