Add extra error info to status description in query log and Cypher Shell ()
Append the position and relevant query line with caret to the
status description of the last GqlStatusObject cause of errors in the
query log and Cypher Shell. This information is part of the legacy error
message and has been requested to be part of the new error system as
well.