Introduce Vector to Query API Plain JSON ()
Plain JSON results in Query API are meant to be human readable. In this
case, the vector is be defined as `Array<Number>`. This is the most
straightforward representation of the `VECTOR `and it aligns well with
the representation in other providers.

The downside of this representation is that it is totally
indistinguishable from Lists and it can be written back as List.
However, this problem is already present in other types which are
returned as strings such as the Temporal types.

The changes for Plain JSON are incremental. Those might not be
considered breaking changes so we don’t need versioning in this case.