Fix bug in SchemaCache returning old SchemaDescriptorSupplier ()
The `SchemaCache` keeps "sets" of all of the schema descriptor
suppliers, such as index and constraint definitions, and their partial
matches. It was found that when updating these sets, the updated value
was discarded.
This occurs when we make an index-backed constraint; the backing index
is created first to validate the constraint, and then updated later once
the constraint has been finalised, with the owning ID. This owning ID
was effectively discarded in the eyes of these lookups, as this
information is not used in the `hashCode` and `equals` of the
definitions.

The core of the problem is the implementation of the `HashSet<E>`. It is
based upon `HashMap<E,Object>`, and uses the key of the map as the
source of truth and uses a dummy value sington `PRESENT` (simply `new
Object()`) as the value for all entries in the map.

The `Set::add` calls the `Map::put` method, which will always replace
the value and return the old one. In this instance, that would simply
replace `PRESENT` with `PRESENT`, and nothing will change the key, which
is what we are using for the entry of the set.

Out of the solutions tried, we chose a set-like class similar to
`HashSet<E>`, but using a `HashMap<E,E>` such that the _latest_ version
is the _value_.

---------

Co-authored-by: emeliebondesson <emelie.bondesson@neo4j.com>