Fail early with vector types from input source if unsupported ()
Not all storage engines support all Value types; i.e., Record formats do
not support the vector types.

Currently the `BatchImporter` will simple read the agnostic `Input`, and
fail on the first write of an unknown type. If `Input` can determine at
least some of the types it contains, then it is possible to handle
earlier. This could be in the form of a warning (such as in store copy)
or failing the import or migration before work is done.

CSV and Parquet files contain this information in their respective
headers, and can provide a more complete set; however, we do not track
this information in storage engines.

What we can do is check if vector store files exist, and has data, to
denote that vector data exists; thus should fail early on record
formats.

For now as vectors are the only nonsupported type, to keep things simple
we are directly asking if the input contains vectors. Future types may
warrent a more general approach of estimating value types from their
input.

---------

Co-authored-by: Satia Herfert <satiaherfert@gmx.de>