Fix wrong error message for vectors unsupported ()
When trying to store vectors in record format, we would get the exact
format name (e.g. "aligned", "standard") from the database config for
the error message.

The database config is not guaranteed to contain the actual format of
the running database though. Specifically, in the following scenarios,
it will not.

* overwrite a `block` DB with an `aligned` DB with `neo4j-admin database
load --from-path=... neo4j --overwrite-destination=true`
* migrate a `block` DB to `aligned`

The reason is that config.get(GraphDatabaseSettings.db_format) on that
DB will actually give "block", since we get that from the
`creation_store_format` property of the database node in the system db.

This PR fixes it so that we get the format name from `RecordFormats`
instead.