Block importer caps memory usage of index population ()
By having the ability for index population threads to internally share
data buffers if it's calculated that all all buffers together would go
beyond a certain percentage of max heap memory.

Threads sharing these buffers have some overhead when adding data to
them, but also reduces parallelism when merging and sorting this data on
"scanCompleted", but at least the import can complete instead of going
nuts with memory usage for a high amount of threads, indexes and (for
SPD) number of shards.