Adding function lookup/checking methods to the ProcedureView  ()
This PR adds `ProcedureView.functionSignature()` as a functions
counterpart to `ProcedureView.procedureSignature()`.

To support this, it adds `ProcedureRegistry.functionSignatureFromId()`.

It also adds `ProcedureView.isFunctionInstanceOf()` which checks whether
the id of a function resolves to a particular type of
`CallableUserFunction`.

The motivation for adding these comes from ABAC built-in function
evaluation where we need to be able to detect when a built-in function
has been shadowed by a user-defined one during the deprecation phase
described in [CIP
230](https://docs.google.com/document/d/1G_tzfTHyl_HZ0MKTUWsAcJi8aDu9VxXjgo-ZXtc52Lk/edit?tab=t.0#heading=h.kdb4jt9vcp34).