RemoteBatchProperties: Fix value mapping for per-row-arguments ()
This pr will make sure that all values in variable parameters for
`RemoteBatchProperties` are mapped to Java driver compatible types.

Previously we used an existing `DefaultValueMapper` which maps Neo4j
values to plain Java types, if such a type exists. This works well for
simple types like `Longs`, `Strings` etc, but for `Point` and other more
complex values we must convert to a type provided by the Java driver.
The `SpdParameterValueMapper` will take care of any spd specific value
conversion.

For context, the reason we map parameter values from runtime is to get
the advantage of parallel runtime when constructing batches. We evaluate
and map the per-row arguments when we add morsels to a new batch, which
in the parallel runtime can happen concurrently. This is opposed to when
we execute a batch where we would have to evaluate each row of the
complete batch single threaded just before doing the remote call.