Distributed create database ()
Introduces a new bootstrapping mechanism where empty stores are
generated in a singe location and replicated through Raft.

Previously all bootstrappers where allowed to create an empty store and
it was relied that they would create identical empty stores. This was
achieved by deleting the log and resetting the checksum. This must
change. Firstly because there is no real guarantee that the empty seeds
generated actually will provide the same ids. Secondly because SPD has
property shards that runs as secondaries, these currently don'y have a
bootstrapping mechanisms to reset the checksums (hence lookup indexes
have been created manually for SPD until now) and thirdly, when merging
the logs it is not possible to use a tx log to bootstrap the raft
mechanics since there is only one log.

The change introduces a SeedStoreEntry in Raft. When a new database is
created it must first replicate a SeedStoreEntry before it can start
replicating transactions. This is part of the lifecycle. For a new empty
database all bootstrappers may attempt to do this. But due to the nature
of how the Raft log imposes a global order, only one will be seen as the
first succesful seed and therfor will be used for all members. For
databases that are seeded, a seed entry is still replicated with the
StoreId of the seeded entry. This Store Id should match all members
otherwise the seeding has not correctly been done and the database will
fail to start.

For databases created on a version that supports seed entries, Raft
bootstrapping will se a NOT_YET_SEEDED state. The distributed seeder
will hold the lifecycle start until that state has been updated, either
by itself replicating a seed or by someone else.

If a database is created on an older version, for example during a roll
before the new version is used, then it will set a state that no seeding
should be done. They will therefor skip the seeding phase and its
lifecycle will behave like it has until today, using the legacy
bootstrapper. After rolling past the next LTS this legacy bootstrapping
logic can be removed.

NOTE: The system database still uses legacy seeding. It will be fixed in
a follow up PR, but this does not block SPD since the system db is not
sharded.

---------

Co-authored-by: Conor Watson <conor.watson@neo4j.com>