Don't allow truncation of corrupt tx logs in catchup. ()
There's a setting `internal.dbms.tx_log.fail_on_corrupted_log_files`
that allows recovery to truncate corrupted parts. This setting is
enabled for all Aura databases. This means that if there's a corruption
in the pulled tx log, the recovery after will get rid of things that we
assume exists. Especially when pulling logs with a required tx, we have
to be sure that the kernel state machine has applied up to a certain tx.
Otherwise, this can result in `CommandApplicationProcess` starting to
apply raft entries from way in the future, thinking that the kernel
state machine is up-to-date, causing corruption in the store.