Remove scope from dependencyManagement in top-level pom ()
If a module A depends on a module B with compile scope, and module B
depends on module C with compile scope also, by default module A
transitively depends on module C with compile scope.
This isn't true if in module A, module C is listed in
`dependencyManagement` with test scope. In such a case, module A
transitively depends on module C but the scope is overriden to test.
This is arguably
[counterintuitive](https://stackoverflow.com/questions/15221299/dependencymanagement-and-scope).
It makes it harder to address the warnings coming out of
`dependency:analyze`.
In this PR, I removed the scope information for both `scalatest` and
`mockito-core`. This is part of [Remove scope information from the
dependencyManagement](https://trello.com/c/TNyuxjtx/31-remove-scope-information-from-the-dependencymanagement).