Compress metadata script in backup header ()
There was an arbitrary 1MB limitation on READING the metadata script. So
we actually wrote down what ever size the script was, but refused to
read it - even it was present. This is just silly. Since this limitation
was not really a limitation it is now removed. We assume that whatever
size the script has it will not be too large in a compressed format,
that we need to protect against it. The `BackupMetadataV2` will be still
able to read headers with uncompressed script as the compressed script
uses a different key.