Deduplicate TransactionClockContext ()
Trello card link:
https://trello.com/c/NKcAqixu/12013-remove-duplicate-transactionclockcontext

There were two near-identical definitions for a class of this name; one
in `org.neo4j.kernel.impl.api`, and the other in
`com.neo4j.dbms.database.virtual`.

The duplication happened back in 2022, and appears to have been
deliberate, as described in [this Trello
card](https://trello.com/c/kimSDrDy). I was unable to detect any danger
from consolidating our needs for a TransactionClockContext to the one in
kernel, hence this change.