DWYW: LogicalPlanBuilder refactor ()
This PR refactors the `LogicalPlanBuilder` to remove some of the
filter/string overloads. The goal is that we have one method per logical
plan, and we can provide default arguments.

All of the important changes are in `AbstractLogicalPlanBuilder`. For
`Expression` parameters, there is now a `ToExpression` 'magnet' type
which can either be an `Expression` or a `String`, where a `String` will
be parsed as an expression. Plus some auto-conversions for literals like
`Long` or `Boolean`.