Finalise Raft Upgrade component ()
The objective whit this PR is to make the RaftVersion upgrade logic
ready for release. The new rafted upgrade logic will live side by side
with the current transaction based solution for now.

The reason for a RaftedVersion logic:
1. It will work for read only dbs and dbs that are not processing any
load.
2. It is necessary for changing the Raft log because it must happend on
RaftAppend to maintain identical logs.

**Versions mapping**
This introduces a new object called DatabaseVersion. The DatabaseVersion
contains a mapping for all relevant components needed to write an
EnvelopeLog. This will, when released, replaced KernelVersion from the
DbmsRuntimeVersion, since kernel version is already in the
DatabaseVersion.

**Alternative log header**
The Raft Log will have a byte reflecting the DatabaseVersion instead of
the KernelVersion in its headers so that it can get all correct version
components. An alternative methods for creating a header has been added.

**Ensuring the log has the correct format**
A version change will trigger an event either because a new version has
been applies or because the snapshot has been installed. Currently, this
new version is only used by the Raft log and therefor there is no event
for when a new version has been applied in the state machine from the
log. This will change in the future when the kernel runtime version is
updated using the Rafted upgrade.

**Implementation changes around the cluster code**
In essence the logic has not changed too much, but there has been some
refactoring. However there is one main difference where the code has
shifted from being coupled with kernel version to instead completely
rely on DatabaseVersions instead.

Other changes are:

1. Fix reader to select version based on log header.
2. Fix upgrade scheduler to use timer service, not risk stacking up
jobs, and to ignore starting all together if at latest version.
3. Move recovery logic to the upgrade manager.
4. I've attempted to clean things up a bit and by doing that simplify
the api of the version upgrade class.


**NOTE** This PR does not release the Rafted version, it is preparing it
to be turned on in a following PR. V0 of DatabaseVersion basically means
its not been activated (in lack of better name)

Part of DATA-129