Validate errors before get TxID on QueryAPI tests ()
Test case were trying to access the transaction id for a failed
transaction.
However, in failed transactions the transaction field of the response is
absent.
So, a NullPointerException is thrown hiding the root cause of the issue.

This changes add a validation for verifying if the response is a failure
before trying to access the transaction id.
In case failed, a exception is thrown with message and code so a debug
can be done.