adding abac metadata function ()
This PR adds the dormant function `abac.oidc.user_attribute` in
preparation for the development of ABAC (Attribute-based access
control).

This function is for use in ABAC Auth Rules as a way of accessing a
users OIDC claims at authorization time

For example: 

`CREATE AUTH RULE region1 SET CONDITION
abac.oidc.user_attribute(‘region’) = ‘region1’`

This function cannot be used in any other context. 

The function will show up in the output of `SHOW FUNCTIONS` if the
feature flag is on (i.e. if
`internal.dbms.security.abac_enabled_authorization_providers` is a
non-empty list).

Co-authored-by: Therese Magnusson <scout.therese@gmail.com>