Clean up old entities from HugeEntityStore in MVCC ()
We've so far left deleted root entities in the multi-root GBPTree in
MVCC because the versions must be accessible. This PR makes the root
layer "version-aware", meaning that it keeps track of the last modifying
version, and only changes the value if it's _defined_. So removals, i.e.
_undefined_ values are no-ops basically. However, when the tree gets
defragmented or shrunk, the deleted values will be removed if not
accessible anymore based on the oldest visible transaction.

Note that this PR doesn't touch the data layer. Anything in there will
still be left behind. That will be fixed in a following PR. https://github.com/neo-technology/neo4j/pull/33290