Fix region page reference initialization ()
`regionPageReference` wasn't accounting for the switch bit, thus when
called concurrently with `allocatePage` on "empty" storage it can skip
initial allocation of region and thus return reference to the zero page
of the version storage which doesn't exist yet. Then this reference
could be read and appear not existent, failing loading of the target
snapshot.