Attempt to simplify LogFiles setup
This tries to improve the relationship between logfiles and reading of
logTail to avoid having to create logfiles objects to get the correct providers.
The logfiles read the logtail on start up themselves and use that to
initialize the LogMetadataProvider.

This is a first step, the providers can be simplified more after this

Note that LogFilesBuilder.withDependencies doesn't override anything for the
TransactionLogFilesProviders, the explicit overrides need to be used instead.