Cleanup of mirror database - next take ()
* Remove upstream from DatabaseReferenceImpl.Mirror - nothing needs it,
and we would require to push down enterprise components again to
community.
* Depart from the uri like definition of an upstream,
* Input and output of database creation has now a `UpstreamDescriptor`
as `mirrorDescriptor` instead of a URI.
* Classes `UpstreamDatabase` and `MirrorSupport` are replaced by
`UpstreamDescriptor`. Its lifecycle is for mirror databases:
* It is created from the user input of the `createMirrorDatabase`
procedure.
* It is extended within the write support that is contains the database
id of the upstream (both for local and remote).
* It is written down by
`TopologyGraphWriter.setUpstreamDatabaseForMirror` - has its own
built-in serializing converting it to a string.
* Read in the
`EnterpriseTopologyGraphDbmsModelV5.fetchDatabaseDescriptor` - it parses
itself from the serialized string.
* For non mirror database `UpstreamDescriptor` is simply created from a
`NamedDatabaseId`
* The procedure for creating the mirrors:
* Is moved into its own class into the enterprise assembly module - it
might need to have access to clustering stuff
* Has a new signature: `database` as the name of the mirror, `mirrors`
as the count of copies (still all secondary), `upstream` is now a map
and either it needs to be `{local:"name of upstream"}` or `{remote:"name
of upstream", addresses:["address1:6000", "address2:6000"]}`.
* Local upstreams are not linked "graph" wise between `:Database` nodes.
The reason being the upstream should be an exact database - as in
identified by an id. When recreate this id changes on the `:Database`
node breaking this.
* Determining the database id of the upstream is now at command time for
both local and remote mirrors. This happens in the write support via a
functional interface called `MirrorResolver`.
* This allows that `ReplicaUpstreams` does not need suppliers for
database ids.
* The remote can have now more addresses defined - these will be used in
a round-robin fashion in a `CatchupAddressProvider`. All the addresses
will be checked.