SystemDiagnostics should be more robust to invalid paths and continue ()
One of the windows tests started failing due to a spurious /n in the
environment Path variable, which isn't allowed in Windows file names.
The SystemDiagnostics should be more robust to file names that can't be
converted to a Path and be able to at least log the remaining
information.