synchronize tx starttime with security (authorization) checks ()
Following on from
[this](https://github.com/neo-technology/neo4j/pull/32537) and
[this](https://github.com/neo-technology/neo4j/pull/33047), this PR
proposes setting transaction start time to be the same as the time used
to evaluate authorization.

For places where we do not directly start a transaction, we use the
current time to evaluate authorization.

The two reasons for this change are:
1) To supply a time to the `LoginContext`s `authorize()` and `roles()`
functions in preparation for when authroization can be a function of
time.
2) To make sure that the time used for authorization is the same time
used for transaction start time where applicable.