Keep daemonErr open long enough to print startup exception ()
Trello Card:
https://trello.com/c/QPnchjZR/12277-neo4j-start-cli-error-reporting

`neo4j-admin server start` starts a process (with main class
`NeoAdminBoot`), which in turn spawns a child process (main class
`Neo4jEnterprise`) which handles the database itself. We have a class
called `ErrorGobbler` in the parent process which waits for content from
the `System.err` of the child process.

The problem at hand is that errors in the child process weren't making
their way to the `ErrorGobbler`. This PR addresses that.