When we truncate/delete tx log files we weren't clearing the LogHeaderCache ()
This is really for correctness, because in practice whenever we roll
over into a new log file version and reuse the version number we update
the LogHeaderCache to the newly created header. However, in one support
incident with raw tx pulling I was worried for a moment that the
corruption we saw might have been due to stale header information.