Some improvements to parallel recovery ()
- Fixes an issue where the "schema boundary" shared lock could get an
acquire count that exceeded the limit of ReentrantReadWriteLock. This
could happen for very large transactions (this shared lock could be
acquired once per command)
- Caps number of threads doing parallel recovery to 16. It's some
arbitrary number, but realistically only a couple of threads is enough
and it's definitely the case that for a high thread count in parallel
recovery A LOT of CPU goes to spinning for coordinating the order in
which transactions can be locked and will have a net negative effect if
too many.