Fix bug with inconsistent ordering of UTF-8 String arrays  ()
String arrays properties were ordered inconsistent from Cypher runtime
and from the index. The result is that the same query could return
different results if executed with and without an index.
See added test in `MiscAcceptanceTest`.

The index uses byte-by-byte comparison of the UTF-8 encoded String.
Runtime used Java's standard `String.compareTo`, which is different,
as Java first encodes the Strings as UTF-16 internally.
Runtime used this only for string arrays. For scalar values,
it depends on whether we have a `UTF8StringValue` (byte-by-byte
comparison of the UTF-8 encoded String) or a `StringWrappingStringValue`
(Java's standard `String.compareTo`).

This commit changes it so that in the runtime, string arrays compare the
elements by the same rules as scalar values. In order to do that,
we change `StringArray` to not wrap a `String[]` anymore, but a
`StringValue[]` instead. This creates some overhead of extra
`StringValue`
objects in the cases where we don't have UTF-8 encoded Strings
(e.g. reading the labels of a node), but in the performance crucial
cases
of reading string array properties from an index or a store, we often
have UTF-8 encoded data to begin with.

Note that the inconsistency between `UTF8StringValue` and
`StringWrappingStringValue` is not fixed in this commit. It is therefore
still possible to get inconsistently ordered results, but the Cypher
queries to trigger this are a bit more involved.
There has been a [previous
attempt](https://github.com/neo-technology/neo4j/pull/488/commits/7ae5ce931ae3aed9a67860969341a307d516bf65)
at fixing this, but it got
[reverted](https://github.com/neo-technology/neo4j/pull/488/commits/ca8d4891396d9f570a5db680924891a1d3cd9e78)
in
the same [PR](https://github.com/neo-technology/neo4j/pull/488).
The inconsistent ordering was also
[documented](https://neo4j.com/docs/cypher-manual/3.5/syntax/values/) in
the past, but that was
removed at some point.

More changes:
* Changes RandomValues to use utf8 text values