KRNL-396 Ensure that matching async index drop request is processed when creating new index with reused id ()
There is a "race" between id reuse and `MultiVersionIndexDropController`
when index id can be reused before corresponding async cleanup is
complete. This results in new index being mistakenly dropped by
`MultiVersionIndexDropController`.

This PR adds `maintenance` call after id for index is reserved
which checks `MultiVersionIndexDropController` queue and performs drop
of old index if there is such request.
This is safe to do because id reuse is also guarded by horizon checks,
meaning that is id was reused transaction that deleted it passed the
horizon.